servers/mcp/requirements.txt
cryptography
fastmcp
google-adk
httpx
python-dotenv
uvicorn
